requirements.txt
torch>=2.0.0
snntorch>=0.7.0
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
plotly>=5.14.0
opencv-python>=4.8.0
PyYAML>=6.0
pytest>=7.3.0
jupyter>=1.0.0
tqdm>=4.65.0
scikit-learn>=1.3.0
pandas>=2.0.0
requests>=2.31.0
pillow>=10.0.0
torchvision>=0.15.0
filterpy>=1.4.5
astropy>=5.3.0
h5py>=3.9.0
seaborn>=0.12.0
ipywidgets>=8.0.0
nbformat>=5.9.0
